profiling tests
git-svn-id: https://repository.astro.ex.ac.uk/torus/trunk@4539 25464f21-6099-402c-b30a-640f1d40d38b

+++ torus/diffusion_mod.F90
@@ -879,7 +879,7 @@ end subroutine gaussSeidelSweep
     integer :: subcell
     integer :: i
     real(double) :: tau
-    real :: kappap
+    real(double) :: kappap
     kappap = 0.d0
     do subcell = 1, thisOctal%maxChildren
        if (thisOctal%hasChild(subcell)) then
